requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
requests==2.32.5
urllib3==2.6.3
watchdog==6.0.0
python-dotenv
